group-project-team-7/app/search/pyproject.toml
[project]
name = "search"
version = "0.1.0"
description = ""
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "voyageai",
    "pymongo",
    "langgraph>=0.3,<0.4",
    "langchain>=0.3,<0.4",
    "langchain-openai",
    "langgraph-checkpoint-mongodb==0.1.3",
    "pydantic-settings",
    "langsmith>=0.3,<0.4",
    "python-jose[cryptography]",
    "python-multipart",
    "passlib[bcrypt]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/search"]
